Put Course into separate module and subcomponents into that module
